court room basic implemented, reviw card made and config
