+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/index.scss
@@ -149,6 +149,7 @@
 @import "./hearings/prehearing.scss";
 /* @import "./home"; */
 @import "./home/home.scss";
+@import "./home//components/reviewCard.scss";
 
 .wbh-header-container {
   display: flex !important;
