+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/pages/employee/taskcomponent.scss
@@ -75,6 +75,9 @@
   }
   .label-field-pair {
     width: 50%;
+    .digit-card-label.card-label {
+      font-size: 16px;
+    }
     .digit-dropdown-employee-select-wrap {
       margin-bottom: 0px;
     }
