+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/components/TaskComponent.js
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { Dropdown } from "@egovernments/digit-ui-components";
+import { CardLabel, Dropdown } from "@egovernments/digit-ui-components";
 import { LabelFieldPair } from "@egovernments/digit-ui-react-components";
 import { Loader } from "@egovernments/digit-ui-react-components";
 import { useGetPendingTask } from "../hooks/useGetPendingTask";
@@ -317,6 +317,7 @@ const TasksComponent = ({ taskType, setTaskType, isLitigant, uuid, filingNumber 
       <h2>Your Tasks</h2>
       <div className="task-filters">
         <LabelFieldPair>
+          <CardLabel className={"card-label"}>{`Case Type`}</CardLabel>
           <Dropdown
             option={[{ name: "NIA S138", code: "NIA S138" }]}
             selected={{ name: "NIA S138", code: "NIA S138" }}
@@ -326,6 +327,7 @@ const TasksComponent = ({ taskType, setTaskType, isLitigant, uuid, filingNumber 
           />
         </LabelFieldPair>
         <LabelFieldPair>
+          <CardLabel className={"card-label"}>{`Task Type`}</CardLabel>
           <Dropdown
             option={taskTypes}
             optionKey={"name"}
