+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/configs/HomeConfig.js
@@ -1,8 +1,8 @@
-const defaultSearchValues = {
-  filingNumber: "",
-  caseType: "NIA S138",
-  substage: "",
-};
+import { TabBenchSearchConfig } from "./BenchHomeConfig";
+import { TabCourtRoomSearchConfig } from "./CourtRoomHomeConfig";
+import { TabFSOSearchConfig } from "./FSOHomeConfig";
+import { TabJudgeSearchConfig } from "./JudgeHomeConfig";
+import { TabLitigantSearchConfig } from "./LitigantHomeConfig";
 
 export const CaseWorkflowState = {
   CASE_RE_ASSIGNED: "CASE_RE_ASSIGNED",
@@ -113,1442 +113,6 @@ export const userTypeOptions = [
   },
 ];
 
-export const TabLitigantSearchConfig = {
-  tenantId: "pg",
-  moduleName: "homeLitigantUiConfig",
-  showTab: true,
-  TabSearchConfig: [
-    {
-      label: "All",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [{}],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeLitigantUiConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues,
-            fields: [
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Stage",
-                isMandatory: false,
-                key: "stage",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "substage",
-                  options: subStageOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  style: { maxWidth: "250px", minWidth: "200px", width: "220px" },
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-            ],
-          },
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-                additionalCustomization: true,
-              },
-              {
-                label: "Stage",
-                jsonPath: "substage",
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Filing Date",
-                jsonPath: "filingDate",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-    },
-    {
-      label: "Drafts",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              status: "DRAFT_IN_PROGRESS",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeLitigantUiConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues, // Set default values for search fields
-            fields: [
-              {
-                type: "component",
-                component: "CustomSortComponent",
-                isMandatory: false,
-                disable: false,
-                name: "Last Edited",
-                key: "sortCaseListByDate",
-                sortBy: "lastModifiedTime",
-                showIcon: true,
-                icon: "UpDownArrowIcon",
-                populators: {},
-              },
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Draft Name",
-                jsonPath: "caseTitle",
-                additionalCustomization: true,
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Last Edited",
-                jsonPath: "auditDetails.lastModifiedTime",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-      additionalDetails: {
-        sortBy: "sortCaseListByDate",
-      },
-    },
-    {
-      label: "Closed",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [{ stage: "Post-Trial" }],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeLitigantUiConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues,
-            fields: [
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Outcome",
-                isMandatory: false,
-                key: "outcome",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "outcome",
-                  options: outcomesOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "OutCome",
-                jsonPath: "outcome",
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Filing Date",
-                jsonPath: "filingDate",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-    },
-  ],
-};
-
-export const TabJudgeSearchConfig = {
-  tenantId: "pg",
-  moduleName: "homeJudgeUIConfig",
-  showTab: true,
-  TabSearchConfig: [
-    {
-      label: "All",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              stage: "",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeJudgeUIConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues,
-            fields: [
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Stage",
-                isMandatory: false,
-                key: "stage",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "substage",
-                  options: subStageOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "Stage",
-                jsonPath: "substage",
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Filing Date",
-                jsonPath: "filingDate",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-    },
-    {
-      label: "Ongoing",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              stage: "Trial",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeJudgeUIConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues, // Set default values for search fields
-            fields: [
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Stage",
-                isMandatory: false,
-                key: "stage",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "substage",
-                  options: subStageOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "Stage",
-                jsonPath: "substage",
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Filing Date",
-                jsonPath: "filingDate",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-    },
-    {
-      label: "Registered",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              stage: "Post-Trial",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeJudgeUIConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues,
-            fields: [
-              {
-                type: "component",
-                component: "CustomSortComponent",
-                isMandatory: false,
-                disable: false,
-                name: "Sort by",
-                key: "sortCaseListByDate",
-                sortBy: "createdtime",
-                ascText: "First",
-                descText: "Last",
-                showAdditionalText: true,
-                showIcon: true,
-                icon: "UpDownArrowIcon",
-                populators: {},
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Stage",
-                isMandatory: false,
-                key: "stage",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "substage",
-                  options: subStageOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "Stage",
-                jsonPath: "substage",
-              },
-
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Filing Date",
-                jsonPath: "filingDate",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-        additionalDetails: {
-          sortBy: "sortCaseListByDate",
-        },
-      },
-    },
-    {
-      label: "Closed",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              stage: "Post-Trial",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeJudgeUIConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues,
-            fields: [
-              {
-                type: "component",
-                component: "CustomSortComponent",
-                isMandatory: false,
-                disable: false,
-                name: "Closed:",
-                key: "sortCaseListByDate",
-                sortBy: "createdtime",
-                ascText: "new first",
-                descText: "old first",
-                showAdditionalText: true,
-                showIcon: true,
-                icon: "UpDownArrowIcon",
-                populators: {},
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Outcome",
-                isMandatory: false,
-                key: "outcome",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "outcome",
-                  options: outcomesOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "Outcome",
-                jsonPath: "outcome",
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Filing Date",
-                jsonPath: "filingDate",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-      additionalDetails: {
-        sortBy: "sortCaseListByDate",
-      },
-    },
-  ],
-};
-
-export const TabBenchSearchConfig = {
-  tenantId: "pg",
-  moduleName: "homeJudgeUIConfig",
-  showTab: true,
-  TabSearchConfig: [
-    {
-      label: "Ongoing",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              stage: "Trial",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeJudgeUIConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues, // Set default values for search fields
-            fields: [
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Stage",
-                isMandatory: false,
-                key: "stage",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "substage",
-                  options: subStageOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "Stage",
-                jsonPath: "substage",
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Filing Date",
-                jsonPath: "filingDate",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-    },
-    {
-      label: "Closed",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              stage: "Post-Trial",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeJudgeUIConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues,
-            fields: [
-              {
-                type: "component",
-                component: "CustomSortComponent",
-                isMandatory: false,
-                disable: false,
-                name: "Closed:",
-                key: "sortCaseListByDate",
-                sortBy: "createdtime",
-                ascText: "new first",
-                descText: "old first",
-                showAdditionalText: true,
-                showIcon: true,
-                icon: "UpDownArrowIcon",
-                populators: {},
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Outcome",
-                isMandatory: false,
-                key: "outcome",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "outcome",
-                  options: outcomesOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "Outcome",
-                jsonPath: "outcome",
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Filing Date",
-                jsonPath: "filingDate",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-      additionalDetails: {
-        sortBy: "sortCaseListByDate",
-      },
-    },
-  ],
-};
-
-export const TabFSOSearchConfig = {
-  tenantId: "pg",
-  moduleName: "homeFSOUiConfig",
-  showTab: true,
-  TabSearchConfig: [
-    {
-      label: "All",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              status: "UNDER_SCRUTINY",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeFSOUiConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: {
-              ...defaultSearchValues,
-              sortCaseListByDate: {
-                sortBy: "createdtime",
-                order: "desc",
-              },
-            },
-            fields: [
-              {
-                type: "component",
-                component: "CustomSortComponent",
-                isMandatory: false,
-                disable: false,
-                name: "Filed",
-                key: "sortCaseListByDate",
-                sortBy: "createdtime",
-                ascText: "(old first)",
-                descText: "(new first)",
-                showAdditionalText: true,
-                showIcon: true,
-                icon: "UpDownArrowIcon",
-                populators: {},
-              },
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Status",
-                isMandatory: false,
-                key: "stage",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "substage",
-                  options: subStageOptions,
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-            ],
-          },
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "Stage",
-                jsonPath: "substage",
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Days Since Filing",
-                jsonPath: "auditDetails.createdTime",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-      },
-      additionalDetails: {
-        sortBy: "sortCaseListByDate",
-      },
-    },
-    {
-      label: "Scrutiny Due",
-      type: "search",
-      apiDetails: {
-        serviceName: "/case/v1/_search",
-        requestParam: {},
-        requestBody: {
-          tenantId: "pg",
-          criteria: [
-            {
-              status: "UNDER_SCRUTINY",
-            },
-          ],
-        },
-        masterName: "commonUiConfig",
-        moduleName: "homeFSOUiConfig",
-        minParametersForSearchForm: 0,
-        tableFormJsonPath: "requestBody",
-        filterFormJsonPath: "requestBody",
-        searchFormJsonPath: "requestBody",
-      },
-      sections: {
-        search: {
-          uiConfig: {
-            formClassName: "custom-both-clear-search",
-            primaryLabel: "ES_COMMON_SEARCH",
-            secondaryLabel: "ES_COMMON_CLEAR_SEARCH",
-            minReqFields: 0,
-            defaultValues: defaultSearchValues, // Set default values for search fields
-            fields: [
-              {
-                type: "component",
-                component: "CustomSortComponent",
-                isMandatory: false,
-                disable: false,
-                name: "Filed",
-                key: "sortCaseListByDate",
-                sortBy: "createdtime",
-                ascText: "(new first)",
-                descText: "(old first)",
-                showAdditionalText: true,
-                showIcon: true,
-                icon: "UpDownArrowIcon",
-                populators: {},
-              },
-              {
-                label: "Case Type",
-                isMandatory: false,
-                key: "caseType",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "caseType",
-                  options: ["NIA S138"],
-                  styles: {
-                    maxWidth: "200px",
-                    minWidth: "150px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Status",
-                isMandatory: false,
-                key: "stage",
-                type: "dropdown",
-                disable: false,
-                populators: {
-                  name: "substage",
-                  options: [],
-                  styles: {
-                    maxWidth: "250px",
-                    minWidth: "200px",
-                  },
-                  optionsCustomStyle: {
-                    overflowX: "hidden",
-                  },
-                },
-              },
-              {
-                label: "Case ID",
-                type: "text",
-                isMandatory: false,
-                disable: false,
-                populators: {
-                  name: "filingNumber",
-                  error: "BR_PATTERN_ERR_MSG",
-                  validation: {
-                    pattern: {},
-                    minlength: 2,
-                  },
-                },
-              },
-            ],
-          },
-
-          show: true,
-        },
-        searchResult: {
-          tenantId: Digit.ULBService.getCurrentTenantId(),
-          uiConfig: {
-            columns: [
-              {
-                label: "Case Name",
-                jsonPath: "caseTitle",
-              },
-              {
-                label: "Scrutiny Status",
-                jsonPath: "status",
-                additionalCustomization: true,
-              },
-              {
-                label: "Case ID",
-                jsonPath: "filingNumber",
-              },
-              {
-                label: "Case Type",
-                jsonPath: "",
-                additionalCustomization: true,
-              },
-              {
-                label: "Days Since Filing",
-                jsonPath: "auditDetails.createdTime",
-                additionalCustomization: true,
-              },
-            ],
-
-            enableColumnSort: true,
-            resultsJsonPath: "criteria[0].responseList",
-          },
-          show: true,
-        },
-        additionalDetails: {
-          sortBy: "sortCaseListByDate",
-        },
-      },
-    },
-  ],
-};
-
 export const rolesToConfigMapping = [
   {
     roles: [
@@ -1614,8 +178,22 @@ export const rolesToConfigMapping = [
     },
   },
   {
-    roles: ["BENCH_CLERK"],
+    roles: ["BENCHCLERK_ROLE"],
     config: TabBenchSearchConfig,
+    isCourtOfficer: true,
+    onRowClickRoute: {
+      dependentUrl: "/dristi/admission",
+      urlDependentOn: "status",
+      urlDependentValue: "PENDING_ADMISSION",
+      params: [
+        { key: "filingNumber", value: "filingNumber" },
+        { key: "caseId", value: "id" },
+      ],
+    },
+  },
+  {
+    roles: ["COURT_ROOM"],
+    config: TabCourtRoomSearchConfig,
     isCourtOfficer: true,
     onRowClickRoute: {
       dependentUrl: "/dristi/admission",
