+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/configs/UICustomizations.js
@@ -128,7 +128,7 @@ export const UICustomizations = {
         case "Case Name":
           return (
             <span className="case-name-on-hover">
-              {row?.caseTitle ? (row?.caseTitle?.trim().endsWith("vs") ? `${row?.caseTitle} _______` : row?.caseTitle) : "CASE_UNTITLED"}
+              {row?.caseTitle ? (row?.caseTitle?.trim().endsWith("vs") ? `${row?.caseTitle} _______` : row?.caseTitle) : t("CASE_UNTITLED")}
             </span>
           );
         case "Case Type":
@@ -224,7 +224,7 @@ export const UICustomizations = {
           const formattedCreatedAt = new Date(createdAt.getFullYear(), createdAt.getMonth(), createdAt.getDate());
           const differenceInTime = formattedToday.getTime() - formattedCreatedAt.getTime();
           const differenceInDays = Math.ceil(differenceInTime / (1000 * 3600 * 24));
-          return <span style={{ color: differenceInDays > 30 && "#9E400A", fontWeight: differenceInDays > 30 ? 500 : 400 }}>{differenceInDays}</span>;
+          return <span style={{ color: differenceInDays > 2 && "#9E400A", fontWeight: differenceInDays > 2 ? 500 : 400 }}>{differenceInDays}</span>;
         default:
           return t("ES_COMMON_NA");
       }
